+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Prop `hoverImageLabel` to `product-summary-image`.
+- Created function to normalize product-summary and use it uniformly in other components.
 
 ## [2.31.0] - 2019-08-01
 ### Added